Friend notifications

+++ app/Http/Controllers/ReactionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Events\FriendshipSent;
 use App\Events\PostReactedAction;
 use App\Events\PostReaction;
+use App\Models\Notification;
 use Illuminate\Http\Request;
 use Validator;
 use App\Models\Post;
@@ -12,6 +13,9 @@ use App\Models\Reaction;
 use Exception;
 use Illuminate\Support\Facades\Log;
 use Tymon\JWTAuth\Facades\JWTAuth;
+
+use function App\Providers\notifyNotificationRecieved;
+use function App\Providers\notifyReaction;
 
 class ReactionController extends Controller
 {
@@ -42,16 +46,22 @@ class ReactionController extends Controller
             return response()->json(['error' => $validator->errors()->first()], 422);
         }
 
+        $post = Post::find(request()->post_id);
+
+        if($post === null){
+            return response()->json(['error' => 'Post not found'], 404);
+        }
+
         $reacted = Reaction::where([
-            'user_id' => $userId,
-            'post_id' => request()->post_id
-            ])->first();
+                        'user_id' => $userId,
+                        'post_id' => request()->post_id
+                    ])->first();
 
         if($reacted !== null){
             // handle change
             if($reacted->reaction_id !== request()->reaction_id){
                 $reacted->load('emotion', 'user');
-                broadcast(new PostReactedAction($reacted, 'removed'));
+                notifyReaction($reacted, 'removed');
                 $reaction = $this->react(request()->post_id, $userId, $fields);
                 $reacted->delete();
                 return response()->json(['msg' => 'Reacted', 'data' => $reaction], 200);
@@ -59,13 +69,57 @@ class ReactionController extends Controller
             //handle delete
             $reaction = $reacted;
             $reaction->load('emotion', 'user');
-            broadcast(new PostReactedAction($reaction, 'removed'));
+            notifyReaction($reacted, 'removed');
             $reacted->delete();
+
+
+            Notification::where('post_id', request()->post_id)
+                ->where('creator', $userId)
+                ->where('type', 'reaction')
+                ->delete();
+
+
             return response()->json(['msg' => 'Removed reaction', 'data' => null], 200);
         }
 
         $reaction = $this->react(request()->post_id, $userId, $fields);
         $reaction->load('emotion', 'user');
+
+        if($post->owner !== $post->creator){
+            if($post->owner !== $userId){
+
+                notifyNotificationRecieved(
+                    'Reacted to your post',
+                    $post->owner,
+                    $userId,
+                    $post->id,
+                    'reaction',
+                );
+            }
+
+            if($post->creator !== $userId){
+
+                notifyNotificationRecieved(
+                    'Reacted to your post',
+                    $post->creator,
+                    $userId,
+                    $post->id,
+                    'reaction',
+                );
+            }
+
+        }else{
+            if($post->owner !== $userId){
+
+                notifyNotificationRecieved(
+                    'Reacted to your post',
+                    $post->owner,
+                    $userId,
+                    $post->id,
+                    'reaction',
+                );
+            }
+        }
 
         return response()->json(['msg' => 'Reacted', 'data' => $reaction], 200);
     }
@@ -73,6 +127,7 @@ class ReactionController extends Controller
     public function react($id, $userId, $fields)
     {
         $post = Post::find($id);
+
         if($post === null){
             return response()->json(['error' => 'Post not found'], 422);
         }
@@ -82,8 +137,8 @@ class ReactionController extends Controller
         $reaction = Reaction::create($fields);
         $reaction->load('emotion', 'user');
 
-        broadcast(new PostReactedAction($reaction, 'add'));
-        // dd(1);
+        notifyReaction($reaction, 'add');
+
         return $reaction;
 
     }
